# Linesweep for polygon validation (ring intersection)