There is a significant performance improvement for polygons that have many interior rings.
```
benchmark                                          old ns/op       new ns/op     delta
BenchmarkPolygonSingleRingValidation/n=10          7356            7608          +3.43%
BenchmarkPolygonSingleRingValidation/n=100         80764           82238         +1.83%
BenchmarkPolygonSingleRingValidation/n=1000        934011          898161        -3.84%
BenchmarkPolygonSingleRingValidation/n=10000       10582933        10235399      -3.28%
BenchmarkPolygonMultipleRingsValidation/n=4        66580           64021         -3.84%
BenchmarkPolygonMultipleRingsValidation/n=36       3912332         774824        -80.20%
BenchmarkPolygonMultipleRingsValidation/n=400      326315405       22045400      -93.24%
BenchmarkPolygonMultipleRingsValidation/n=4096     39136093164     896526026     -97.71%

benchmark                                          old allocs     new allocs     delta
BenchmarkPolygonSingleRingValidation/n=10          99             104            +5.05%
BenchmarkPolygonSingleRingValidation/n=100         900            905            +0.56%
BenchmarkPolygonSingleRingValidation/n=1000        8907           8912           +0.06%
BenchmarkPolygonSingleRingValidation/n=10000       89704          89709          +0.01%
BenchmarkPolygonMultipleRingsValidation/n=4        1046           797            -23.80%
BenchmarkPolygonMultipleRingsValidation/n=36       47042          12338          -73.77%
BenchmarkPolygonMultipleRingsValidation/n=400      5223220        352232         -93.26%
BenchmarkPolygonMultipleRingsValidation/n=4096     539280124      10545142       -98.04%

benchmark                                          old bytes       new bytes     delta
BenchmarkPolygonSingleRingValidation/n=10          4300            4526          +5.26%
BenchmarkPolygonSingleRingValidation/n=100         39632           39858         +0.57%
BenchmarkPolygonSingleRingValidation/n=1000        379272          379498        +0.06%
BenchmarkPolygonSingleRingValidation/n=10000       4910661         4910886       +0.00%
BenchmarkPolygonMultipleRingsValidation/n=4        33864           26930         -20.48%
BenchmarkPolygonMultipleRingsValidation/n=36       1483642         398482        -73.14%
BenchmarkPolygonMultipleRingsValidation/n=400      164503972       11134442      -93.23%
BenchmarkPolygonMultipleRingsValidation/n=4096     16995780992     330619032     -98.05%
```